+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/__init__.py
@@ -29,6 +29,8 @@ from .au_tfn_recognizer import AuTfnRecognizer
 from .au_medicare_recognizer import AuMedicareRecognizer
 from .it_driver_license_recognizer import ItDriverLicenseRecognizer
 from .it_fiscal_code_recognizer import ItFiscalCodeRecognizer
+from .it_identity_card_recognizer import ItIdentityCardRecognizer
+from .it_passport_recognizer import ItPassportRecognizer
 
 NLP_RECOGNIZERS = {
     "spacy": SpacyRecognizer,
@@ -65,4 +67,6 @@ __all__ = [
     "TransformersRecognizer",
     "ItDriverLicenseRecognizer",
     "ItFiscalCodeRecognizer",
+    "ItIdentityCardRecognizer",
+    "ItPassportRecognizer"
 ]